通过precompile阶段的代码生成机制来自动生成BinaryWordType的编解码器

+++ nop-record/src/main/java/io/nop/record/input/ByteBufferRecordBinaryInput.java
@@ -246,8 +246,8 @@ public class ByteBufferRecordBinaryInput implements IRecordBinaryInput {
     //region Unsigned
 
     @Override
-    public int readU1() {
-        return bb.get() & 0xff;
+    public short readU1() {
+        return (short)(bb.get() & 0xff);
     }
 
     //region Big-endian
