+++ nop-record/src/main/java/io/nop/record/netty/ByteBufRecordBinaryInput.java
@@ -251,7 +251,7 @@ public class ByteBufRecordBinaryInput implements IRecordBinaryInput {
     //region Unsigned
 
     @Override
-    public int readU1() {
+    public short readU1() {
         return bb.readUnsignedByte();
     }
